+++ src/main/java/com/ats/adminpanel/model/grngvn/GetGrnGvnDetails.java
@@ -27,7 +27,7 @@ String itemHsncd;
 	private float itemMrp;
 
 	
-	private int grnGvnQty;
+	private float grnGvnQty;
 
 	
 	private float grnGvnAmt;
@@ -88,7 +88,7 @@ String itemHsncd;
 	private int delStatus;
 	
 	
-	private int grnGvnQtyAuto;
+	private float grnGvnQtyAuto;
 	
 	private String itemName;
 	
@@ -146,9 +146,9 @@ String itemHsncd;
 		
 		//23 FEB new Fields to handle qty variation between entry(insert) and dispatch
 		
-				int aprQtyGate;
-				int aprQtyStore;
-				int aprQtyAcc;
+		float aprQtyGate;
+		float aprQtyStore;
+		float aprQtyAcc;
 				
 				
 				float aprTaxableAmt;
@@ -270,11 +270,11 @@ String itemHsncd;
 		this.itemMrp = itemMrp;
 	}
 
-	public int getGrnGvnQty() {
+	public float getGrnGvnQty() {
 		return grnGvnQty;
 	}
 
-	public void setGrnGvnQty(int grnGvnQty) {
+	public void setGrnGvnQty(float grnGvnQty) {
 		this.grnGvnQty = grnGvnQty;
 	}
 
@@ -430,11 +430,11 @@ String itemHsncd;
 		this.delStatus = delStatus;
 	}
 
-	public int getGrnGvnQtyAuto() {
+	public float getGrnGvnQtyAuto() {
 		return grnGvnQtyAuto;
 	}
 
-	public void setGrnGvnQtyAuto(int grnGvnQtyAuto) {
+	public void setGrnGvnQtyAuto(float grnGvnQtyAuto) {
 		this.grnGvnQtyAuto = grnGvnQtyAuto;
 	}
 
@@ -534,27 +534,27 @@ String itemHsncd;
 		this.grnGvnHeaderId = grnGvnHeaderId;
 	}
 
-	public int getAprQtyGate() {
+	public float getAprQtyGate() {
 		return aprQtyGate;
 	}
 
-	public void setAprQtyGate(int aprQtyGate) {
+	public void setAprQtyGate(float aprQtyGate) {
 		this.aprQtyGate = aprQtyGate;
 	}
 
-	public int getAprQtyStore() {
+	public float getAprQtyStore() {
 		return aprQtyStore;
 	}
 
-	public void setAprQtyStore(int aprQtyStore) {
+	public void setAprQtyStore(float aprQtyStore) {
 		this.aprQtyStore = aprQtyStore;
 	}
 
-	public int getAprQtyAcc() {
+	public float getAprQtyAcc() {
 		return aprQtyAcc;
 	}
 
-	public void setAprQtyAcc(int aprQtyAcc) {
+	public void setAprQtyAcc(float aprQtyAcc) {
 		this.aprQtyAcc = aprQtyAcc;
 	}
