+++ src/main/java/com/ats/adminpanel/model/grngvn/TempGrnGvnBeanUp.java
@@ -28,11 +28,11 @@ public class TempGrnGvnBeanUp {
 	private String approvedRemarkAcc;
 	
 	//23 FEB new Fields to handle qty variation between entry(insert) and dispatch
-		int aprQtyGate;
+		float aprQtyGate;
 		
-		int aprQtyStore;
+		float aprQtyStore;
 		
-		int aprQtyAcc;
+		float aprQtyAcc;
 		
 		float aprTaxableAmt;
 		
@@ -139,27 +139,27 @@ public class TempGrnGvnBeanUp {
 			this.approvedRemarkAcc = approvedRemarkAcc;
 		}
 
-		public int getAprQtyGate() {
+		public float getAprQtyGate() {
 			return aprQtyGate;
 		}
 
-		public void setAprQtyGate(int aprQtyGate) {
+		public void setAprQtyGate(float aprQtyGate) {
 			this.aprQtyGate = aprQtyGate;
 		}
 
-		public int getAprQtyStore() {
+		public float getAprQtyStore() {
 			return aprQtyStore;
 		}
 
-		public void setAprQtyStore(int aprQtyStore) {
+		public void setAprQtyStore(float aprQtyStore) {
 			this.aprQtyStore = aprQtyStore;
 		}
 
-		public int getAprQtyAcc() {
+		public float getAprQtyAcc() {
 			return aprQtyAcc;
 		}
 
-		public void setAprQtyAcc(int aprQtyAcc) {
+		public void setAprQtyAcc(float aprQtyAcc) {
 			this.aprQtyAcc = aprQtyAcc;
 		}
